% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{addmissingcolumns}
\alias{addmissingcolumns}
\title{Add missing columns}
\usage{
addmissingcolumns(df, cols)
}
\arguments{
\item{df}{a data table}

\item{cols}{a character vector of all possible columns}
}
\value{
a data table
}
\description{
Add missing columns
}
\details{
This function \bold{modifies by reference}.
This function compares a data tables columns to a character vector of all
possible column names. Any columns not present in \code{df} that are present in \code{cols}
are added to \code{df} with a value of \code{NA_character}.
}
\examples{
x <- data.table::data.table(a = c(1,2,3))
cols <- c("a","b","c")
nsqipr:::addmissingcolumns(x, cols)
x
}
\keyword{internal}
