% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{bmi}
\alias{bmi}
\title{Calculate BMI}
\usage{
bmi(height, weight)
}
\arguments{
\item{height}{a numeric vector}

\item{weight}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Calculate BMI
}
\examples{
bmi(72, 180)

}
