% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcon.R
\name{build_tmp_dir}
\alias{build_tmp_dir}
\title{Create a temporary directory}
\usage{
build_tmp_dir()
}
\value{
a path to a temporary directory
}
\description{
References a temporary directory where .exe files can be unzipped and .csv files can be created.
Creates the directory if it does not already exist.
}
\keyword{internal}
