% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{check_comaneurograft}
\alias{check_comaneurograft}
\title{Remove coma, neuro deficit, and graft columns after 2010}
\usage{
check_comaneurograft(df)
}
\arguments{
\item{df}{a data.table from which to remove the coma, neuro deficit, and graft columns}
}
\value{
a data.table
}
\description{
Remove coma, neuro deficit, and graft columns after 2010
}
\details{
According to NSQIP, Graft failure, Coma, and Peripheral Nerve Injury should not be
considered accurate for any PUF after 2010 (see the \link[https://www.facs.org/quality-programs/acs-nsqip/participant-use]{NSQIP} website).
}
\examples{
x <- data.table::data.table(coma = c(TRUE, TRUE, FALSE), cnscoma = c(TRUE, TRUE, FALSE), ncnscoma = c(1,2,3), dcnscoma = c(1,2,3),
                            neurodef = c(TRUE, TRUE, FALSE), nneurodef = c(1,2,3), dneurodef = c(1,2,3),
                            othgrafl = c(TRUE, TRUE, FALSE), nothgrafl = c(1,2,3), dothgrafl = c(1,2,3),
                            distraction = c("Test","test","test"))
conv_pufyear(x, "acs_nsqip_puf12.txt")
check_comagraftpn(x)

x <- data.table::data.table(coma = c(TRUE, TRUE, FALSE), cnscoma = c(TRUE, TRUE, FALSE), ncnscoma = c(1,2,3), dcnscoma = c(1,2,3),
                            neurodef = c(TRUE, TRUE, FALSE), nneurodef = c(1,2,3), dneurodef = c(1,2,3),
                            othgrafl = c(TRUE, TRUE, FALSE), nothgrafl = c(1,2,3), dothgrafl = c(1,2,3),
                            distraction = c("Test","test","test"))
conv_pufyear(x, "acs_nsqip_puf10.txt")
check_comagraftpn(x)

}
\keyword{internal}
