% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-long.R
\name{check_if_in}
\alias{check_if_in}
\title{Check if columns in data frame}
\usage{
check_if_in(cols, df)
}
\arguments{
\item{cols}{a character vector of column names}

\item{df}{a data.table}
}
\value{
a logical vector
}
\description{
Conducts a quick case-insensitive search using \code{\%qsin\%} of provided column names
and returns TRUE if any of these columns are present in the provided data frame.
}
\seealso{
\%qsin\%
}
\keyword{internal}
