% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_separator}
\alias{check_separator}
\title{Checks a path name for a final back- or forward slash.}
\usage{
check_separator(path)
}
\arguments{
\item{path}{a directory path}
}
\description{
Checks a path name for a final back- or forward slash.
}
\keyword{internal}
