% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindo.R
\name{checkAnyDead}
\alias{checkAnyDead}
\title{Check if any trues exist row-wise in a logical matrix.}
\usage{
checkAnyDead(...)
}
\arguments{
\item{...}{a list of vectors of any type or a data frame}
}
\value{
a logical vector. TRUE if any NA values in row, FALSE if no NA values in row.
}
\description{
Check if any trues exist row-wise in a logical matrix.
}
\keyword{internal}
