% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindo.R
\name{checkAnyTrue}
\alias{checkAnyTrue}
\title{Check if any trues exist row-wise in a logical matrix.}
\usage{
checkAnyTrue(...)
}
\arguments{
\item{...}{a list of logical vectors or a data frame}
}
\value{
a logical vector. TRUE if any trues, FALSE if no trues.
}
\description{
Check if any trues exist row-wise in a logical matrix.
}
\keyword{internal}
