% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindo.R
\name{checkTrue}
\alias{checkTrue}
\title{Check if a boolean is TRUE and not NA}
\usage{
checkTrue(bool)
}
\arguments{
\item{bool}{a logical vector}
}
\value{
a logical vector
}
\description{
This represent a key assumption about the Dindo classification sorting. If a complication is missing (NA), it is considered FALSE
for the purposes of the Dindo classification.
}
\keyword{internal}
