% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{coalesce_cols}
\alias{coalesce_cols}
\title{Coalesce two columns}
\usage{
coalesce_cols(df, coalesce_in_cols, coalesce_out_cols)
}
\arguments{
\item{df}{a data table with the desired columns to coalesce}

\item{coalesce_in_cols}{a character vector of column names representing the columns to be coalesced into and that take precedence}

\item{coalesce_out_cols}{a character vector of column names representing the columns to be coalesced out of}
}
\value{
a data table
}
\description{
Coalesce two columns
}
\details{
This function \bold{modifies by reference}.
The \code{coalesce_in_cols} columns will take precedence over the values in the \code{coalesce_out_cols} columns if both
columns contain data. The data is coalesced into the \code{coalesce_in_cols} so that the data in the \code{coalesce_out_cols} is
not altered. The \code{coalesce_out_cols} columns are not removed.

If \code{df} does not contain any \code{coalesce_out_cols} columns, no changes to \code{df} are made. If \code{df}
contains \code{coalesce_out_cols} columns but no \code{coalesce_in_cols}, the \code{coalesce_out_cols} columns are simply
renamed to their counterpart \code{coalesce_in_cols} columns. For this reason, \code{coalesce_in_cols} and
\code{coalesce_out_cols} must be the same length and matching indices between the two vectors represent respective columns
to be coalesced into and out of.
}
\examples{
x <- data.table::data.table(a = c(NA, TRUE, FALSE), b = c(NA, TRUE, FALSE),
c = c(NA, TRUE, FALSE), d = c(NA, TRUE, FALSE),
A = c(TRUE, NA, TRUE), B = c(TRUE, NA, TRUE))
coalesce_in_cols <- c("A","B","C","D")
coalesce_out_cols <- c("a","b","c","d")
coalesce_cols(x, coalesce_in_cols, coalesce_out_cols)
x

}
\keyword{internal}
