% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cols.R
\name{collect_column_names}
\alias{collect_column_names}
\alias{get_headers}
\title{Collect all column names}
\usage{
collect_column_names(files)

get_headers(file)
}
\arguments{
\item{files}{list of files from which to collect column names}

\item{file}{a file to collect names from}
}
\value{
a character vector

a character vector
}
\description{
Collect all column names

Collect column names
}
\details{
Reads the first line of each input file to collect column names.
Creates a character vector and selects only unique entries after converting
to lower case. This results in a vector of all possible column names.

Reads the first line of an input file to collect column names.
Creates a character vector and converts to lower case.
}
\section{Functions}{
\itemize{
\item \code{get_headers}: 
}}

\keyword{internal}
