% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{colorder}
\alias{colorder}
\title{Set data table columns to a specified order}
\usage{
colorder(df, col_order)
}
\arguments{
\item{df}{a data table to be ordered}

\item{col_order}{a character vector of column names in the desired order}
}
\value{
a data table
}
\description{
Set data table columns to a specified order
}
\details{
This function \bold{modifies by reference}.
}
\examples{
x <- data.table::data.table(c = c(1,2,3), b = c(1,2,3), a = c(1,2,3))
col_order <- c("a","b","c")
colorder(x, col_order)
identical(names(x), col_order)

}
\keyword{internal}
