% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{conv_}
\alias{conv_}
\title{Apply a given function to specified columns in a data table}
\usage{
conv_(df, cols, f, ..., newcol)
}
\arguments{
\item{df}{a data table}

\item{cols}{a character vector of column names to which the function \code{f} is applied}

\item{f}{a function that applies to each column.}

\item{...}{arguments to pass to function \code{f}}

\item{newcol}{a character vector of new column names to which the output of \code{f} will be applied.}
}
\description{
Apply a given function to specified columns in a data table
}
\details{
This function \bold{modifies by reference}.
}
\examples{
x <- data.table::data.table(x = rep("APPLE", 10), y = rep("BANANA", 10), z = rep("CHERRY", 10))
conv_(x, c("x","y"), tolower)
x

x <- data.table::data.table(x = rep("APPLE", 10), y = rep("BANANA", 10), z = rep("CHERRY", 10))
conv_(x, c("x","y"), tolower, c("X","Y"))
x

x <- data.table::data.table(x = rep("APPLE", 10), y = rep("BANANA", 10), z = rep("CHERRY", 10))
conv_(x, "x", paste, "JUICE", sep = " ", newcol = "drink")

}
