% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_aaa.R
\name{conv_aaa_cols}
\alias{conv_aaa_cols}
\title{Convert targeted abdominal aortic aneurysm columns}
\usage{
conv_aaa_cols(df, filename)
}
\arguments{
\item{df}{a data table to be cleaned}

\item{filename}{the name of the file from which the data table has been read in}
}
\description{
Convert targeted abdominal aortic aneurysm columns
}
\details{
If the file being processed is a targeted abdominal aortic aneurysm (AAA) dataset,
it will be processed by this function. This function determines how data cleaning steps specific
to targeted AAA files should proceed.
}
\keyword{internal}
