% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_aaa.R
\name{conv_aaa_iculos}
\alias{conv_aaa_iculos}
\title{Convert targeted AAA ICU length of stay to integer}
\usage{
conv_aaa_iculos(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
an integer vector
}
\description{
Convert targeted AAA ICU length of stay to integer
}
\details{
NSQIP encodes anyone staying 30 days or more as "30 or more".
This converts all "30 or more" to 30. If given NA, will return NA.
}
\examples{
nsqipr:::conv_aaa_iculos(c("14","2","30 or more",NA))

}
\keyword{internal}
