% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{conv_acs_cols}
\alias{conv_acs_cols}
\title{Convert ACS NSQIP PUF columns}
\usage{
conv_acs_cols(df, filename)
}
\arguments{
\item{df}{a data table to be cleaned}

\item{filename}{the name of the file from which the data table has been read in}
}
\description{
Convert ACS NSQIP PUF columns
}
\details{
If the file being processed is an ACS NSQIP PUF (i.e., not a targeted data set),
it will be processed by this function. This function determines how data cleaning steps specific
to ACS NSQIP PUF files should proceed.
}
\keyword{internal}
