% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{conv_age}
\alias{conv_age}
\title{Convert age to integer}
\usage{
conv_age(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
an integer vector
}
\description{
Convert age to integer
}
\details{
NSQIP encodes anyone over the age of 90 as "90+". This converts all "90+" to 90.
If given NA, will return NA.
}
\examples{

}
\keyword{internal}
