% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_aie.R
\name{conv_aie_cols}
\alias{conv_aie_cols}
\title{Convert targeted aortoiliac endo columns}
\usage{
conv_aie_cols(df, filename)
}
\arguments{
\item{df}{a data table to be cleaned}

\item{filename}{the name of the file from which the data table has been read in}
}
\description{
Convert targeted aortoiliac endo columns
}
\details{
If the file being processed is a targeted aortoiliac endo dataset,
it will be processed by this function. This function determines how data cleaning steps specific
to targeted aortoiliac endo dataset files should proceed.
}
\keyword{internal}
