% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_aio.R
\name{conv_aio_cols}
\alias{conv_aio_cols}
\title{Convert targeted aortoiliac open columns}
\usage{
conv_aio_cols(df, filename)
}
\arguments{
\item{df}{a data table to be cleaned}

\item{filename}{the name of the file from which the data table has been read in}
}
\description{
Convert targeted aortoiliac open columns
}
\details{
If the file being processed is a targeted aortoiliac open dataset,
it will be processed by this function. This function determines how data cleaning steps specific
to targeted aortoiliac open files should proceed.
}
\keyword{internal}
