% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_app.R
\name{conv_app_cols}
\alias{conv_app_cols}
\title{Convert targeted appendectomy columns}
\usage{
conv_app_cols(df, filename)
}
\arguments{
\item{df}{a data table to be cleaned}

\item{filename}{the name of the file from which the data table has been read in}
}
\description{
Convert targeted appendectomy columns
}
\details{
If the file being processed is a targeted appendectomy data set,
it will be processed by this function. This function determines how data cleaning steps specific
to targeted appendectomy files should proceed.
}
\keyword{internal}
