% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_app.R
\name{conv_app_img_ct}
\alias{conv_app_img_ct}
\title{Parse entries that indicate a CT was performed}
\usage{
conv_app_img_ct(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate a CT was performed
}
\details{
returns TRUE if case-insensitive "ct done-"
is detected in the character vector.
}
\examples{
x <- c("CT not performed/not documented","CT done-result consistent with diagnosis of appendicitis",
"CT done-result not consistent with appendicitis","CT done-result indeterminate; result uncertain", NA)

cbind(x, nsqipr:::conv_app_img_ct(x))

}
\keyword{internal}
