% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_app.R
\name{conv_app_img_ultra}
\alias{conv_app_img_ultra}
\title{Parse entries that indicate an ultrasound was performed}
\usage{
conv_app_img_ultra(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate an ultrasound was performed
}
\details{
returns TRUE if case-insensitive "us done-"
is detected in the character vector.
}
\examples{
x <- c("US done-consistent with appendicitis","US not performed/not documented",
"US done-indeterminate; result uncertain","US done-not consistent with appendicitis", NA)

cbind(x, nsqipr:::conv_app_img_ultra(x))

}
\keyword{internal}
