% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_app.R
\name{conv_app_intraabscess}
\alias{conv_app_intraabscess}
\title{Parse entries that indicate an post-operative intraabdominal abscess}
\usage{
conv_app_intraabscess(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate an post-operative intraabdominal abscess
}
\details{
returns TRUE if case-insensitive "^yes-"
is detected in the character vector.
}
\examples{
x <- c("No diagnosis of postoperative abscess", "Yes-percutaneous drainage",
"Yes-no treatment or intervention", "Yes-IV antibiotics w/out procedural intervention",
"Yes-oral/IM antibiotics w/out procedural intervention",
"Yes-reoperation for surgical drainage, open", "Yes-reoperation for surgical drainage, minimally invasive",
"Yes-transrectal/other endoscopic drainage", NA)

cbind(x, nsqipr:::conv_app_intraabscess(x))

}
\keyword{internal}
