% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_app.R
\name{conv_app_intraabscess_intervention}
\alias{conv_app_intraabscess_intervention}
\title{Parse a column for type of intra-abdominal abscess intervention}
\usage{
conv_app_intraabscess_intervention(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a factor vector
}
\description{
Parse a column for type of intra-abdominal abscess intervention
}
\details{
NSQIP encodes the \code{app_intraabscess} column as having one of multiple interventions.
This function extracts those values from character vectors and factors them.
}
\examples{
x <- c("No diagnosis of postoperative abscess", "Yes-percutaneous drainage",
"Yes-no treatment or intervention", "Yes-IV antibiotics w/out procedural intervention",
"Yes-oral/IM antibiotics w/out procedural intervention",
"Yes-reoperation for surgical drainage, open", "Yes-reoperation for surgical drainage, minimally invasive",
"Yes-transrectal/other endoscopic drainage", NA)

nsqipr:::conv_app_intraabscess_intervention(x)

}
\keyword{internal}
