% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_app.R
\name{conv_app_perforation}
\alias{conv_app_perforation}
\title{Parse entries that indicate a perforation}
\usage{
conv_app_perforation(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate a perforation
}
\details{
returns TRUE if case-insensitive "^perforation only|^perforation and abscess"
is detected in the character vector.
}
\examples{
x <- c("No mention of perforation or abscess", "Perforation and abscess",
"Perforation only", "Abscess only", NA)

cbind(x, nsqipr:::conv_app_perforation(x))

}
\keyword{internal}
