% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_col.R
\name{conv_col_anastomotic}
\alias{conv_col_anastomotic}
\title{Parse entries that indicate the presence of an anastomotic leak}
\usage{
conv_col_anastomotic(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate the presence of an anastomotic leak
}
\details{
returns TRUE if case-insensitive "yes" or "leak'
is detected in the character vector.
}
\examples{
x <- c("No", "Yes-reoperation", "Yes-percutaneous intervention",
"Yes-no intervention required",
"No definitive diagnosis of leak/leak related abscess",
"Leak, treated w/ reoperation", "Leak, treated w/ interventional means",
"Leak, no treatment intervention documented",
"Leak, treated w/ non-interventional/non-operative means", NA)

cbind(x, nsqipr:::conv_col_anastomotic(x))

}
\keyword{internal}
