% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_col.R
\name{conv_col_leak_treatment}
\alias{conv_col_leak_treatment}
\title{Parse a column for type of anastomotic leak intervention}
\usage{
conv_col_leak_treatment(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a factor vector
}
\description{
Parse a column for type of anastomotic leak intervention
}
\details{
NSQIP encodes the \code{col_anastomotic} column as having one of multiple interventions.
This function extracts those values from character vectors and factors them.
}
\examples{
x <- c("No", "Yes-reoperation", "Yes-percutaneous intervention",
"Yes-no intervention required",
"No definitive diagnosis of leak/leak related abscess",
"Leak, treated w/ reoperation", "Leak, treated w/ interventional means",
"Leak, no treatment intervention documented",
"Leak, treated w/ non-interventional/non-operative means", NA)

nsqipr:::conv_col_leak_treatment(x)

}
\keyword{internal}
