% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{conv_date}
\alias{conv_date}
\title{Converts a single year to a date}
\usage{
conv_date(vec)
}
\arguments{
\item{vec}{a character vector containing dates in the format "yyyy"}
}
\value{
a Date vector
}
\description{
Converts a single year to a date
}
\details{
Formats dates via \code{as.Date(vec, "\%Y")}
}
\examples{

x <- c("2000", "1900", "2020", "1950")
nsqipr:::conv_date(x)

}
\keyword{internal}
