% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{conv_factor}
\alias{conv_factor}
\title{Convert factor columns}
\usage{
conv_factor(df, factor_cols)
}
\arguments{
\item{df}{a data table in which to convert factor columns}

\item{factor_cols}{columns to be converted to a factor}
}
\value{
a data table
}
\description{
Convert factor columns
}
\details{
This function \bold{modifies by reference}.
This function checks for which columns to factor by comparing against a character vector called \code{factor_cols}.
The function then fetches a variable from the calling environment with the same name as the column being converted to a factor.
This variable should reference a named list specifying how to rename the levels (see \code{\link[nsqipr:factorpipe]{\%^\%}})
for further details.
}
\examples{

x <- data.table::data.table(foods = c("apple","banana","cherry","steak","chicken","pork"),
                            drinks = c("milk","water","oj","beer","vodka","rum"))
foods <- list(fruit = c("apple", "banana", "cherry"), meat = c("steak","chicken","pork"))
drinks <- list(`non-alcoholic` = c("milk","water","oj"), alcoholic = c("beer","vodka","rum"))
factor_cols <- c("foods", "drinks")
nsqipr:::conv_factor(x, factor_cols)
x

}
\keyword{internal}
