% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_bileleakage}
\alias{conv_hep_bileleakage}
\title{Parse entries that indicate the presence of a bile leak}
\usage{
conv_hep_bileleakage(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate the presence of a bile leak
}
\details{
returns TRUE if case-insensitive "yes"
is detected in the character vector.
}
\examples{
x <- c("No",
"Yes-clinical diagnosis, drain continued on or after POD3",
"Yes-clinical diagnosis, percutaneous drainage performed",
"Yes-clinical diagnosis, reoperation performed",
"Yes-clinical diagnosis, spontaneous wound drainage",
"Yes-persistent drainage, drain continued on or after POD3",
"Yes-persistent drainage, percutaneous drainage performed",
"Yes-persistent drainage, reoperation performed",
NA)

cbind(x, nsqipr:::conv_hep_bileleakage(x))

}
\keyword{internal}
