% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_bileleakage_type}
\alias{conv_hep_bileleakage_type}
\title{Parse a column for type of bile leakage}
\usage{
conv_hep_bileleakage_type(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a factor vector
}
\description{
Parse a column for type of bile leakage
}
\details{
NSQIP encodes the \code{hep_bileleakage} column as either a clinical diagnosis or
persistent drainage. This function extracts those values from character vectors
and factors them.
}
\examples{
x <- c("No",
"Yes-clinical diagnosis, drain continued on or after POD3",
"Yes-clinical diagnosis, percutaneous drainage performed",
"Yes-clinical diagnosis, reoperation performed",
"Yes-clinical diagnosis, spontaneous wound drainage",
"Yes-persistent drainage, drain continued on or after POD3",
"Yes-persistent drainage, percutaneous drainage performed",
"Yes-persistent drainage, reoperation performed",
NA)

nsqipr:::conv_hep_bileleakage_type(x)

}
\keyword{internal}
