% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_cols}
\alias{conv_hep_cols}
\title{Convert targeted hepatectomy columns}
\usage{
conv_hep_cols(df, filename)
}
\arguments{
\item{df}{a data table to be cleaned}

\item{filename}{the name of the file from which the data table has been read in}
}
\description{
Convert targeted hepatectomy columns
}
\details{
If the file being processed is a targeted hepatectomy data set,
it will be processed by this function. This function determines how data cleaning steps specific
to targeted hepatectomy files should proceed.
}
\keyword{internal}
