% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_con_partres}
\alias{conv_hep_con_partres}
\title{Convert concurrent partial resections to integer}
\usage{
conv_hep_con_partres(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
an integer vector
}
\description{
Convert concurrent partial resections to integer
}
\details{
NSQIP encodes anyone with more than 10 concurrent partial resections as "10 or more".
This converts all "10 or more" to 10. If given NA, will return NA.
}
\examples{
nsqipr:::conv_hep_con_partres(c("1","2","10 or more",NA))

}
\keyword{internal}
