% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_hepc}
\alias{conv_hep_hepc}
\title{Parse entries that indicate a Hepatitis C infection}
\usage{
conv_hep_hepc(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate a Hepatitis C infection
}
\details{
returns TRUE if case-insensitive "hepatitis c"
is detected in the character vector.
}
\examples{
x <- c("None","Hepatitis B","Unknown","Hepatitis B and C","Hepatitis C",
"Other", NA)

cbind(x, nsqipr:::conv_hep_hepc(x))

}
\keyword{internal}
