% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_liverfail}
\alias{conv_hep_liverfail}
\title{Parse entries that indicate the presence of liver failure}
\usage{
conv_hep_liverfail(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate the presence of liver failure
}
\details{
returns TRUE if case-insensitive "yes"
is detected in the character vector.
}
\examples{
x <- c("No-does not meet criteria for PHLF",
"Yes-meets criteria for PHLF",
"Yes-PHLF (receiving clotting factors to maintain INR)",
NA)

cbind(x, nsqipr:::conv_hep_liverfail(x))

}
\keyword{internal}
