% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_liverfail_type}
\alias{conv_hep_liverfail_type}
\title{Parse a column for type of liver failure}
\usage{
conv_hep_liverfail_type(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a factor vector
}
\description{
Parse a column for type of liver failure
}
\details{
NSQIP encodes the \code{hep_liverfail} column as either "meets criteria for PHLF" or
"PHFL (receiving clotting factors to maintain INR)". This function extracts those values from character vectors
and factors them.
}
\examples{
x <- c("No-does not meet criteria for PHLF",
"Yes-meets criteria for PHLF",
"Yes-PHLF (receiving clotting factors to maintain INR)",
NA)

nsqipr:::conv_hep_liverfail_type(x)

}
\keyword{internal}
