% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_otherviral}
\alias{conv_hep_otherviral}
\title{Parse entries that indicate a non-Hepatitis viral infection}
\usage{
conv_hep_otherviral(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate a non-Hepatitis viral infection
}
\details{
returns TRUE if case-insensitive "other"
is detected in the character vector.
}
\examples{
x <- c("None","Hepatitis B","Unknown","Hepatitis B and C","Hepatitis C",
"Other", NA)

cbind(x, nsqipr:::conv_hep_otherviral(x))

}
\keyword{internal}
