% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hep.R
\name{conv_hep_sec_numtumors}
\alias{conv_hep_sec_numtumors}
\title{Convert number of secondary tumors to integer}
\usage{
conv_hep_sec_numtumors(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
an integer vector
}
\description{
Convert number of secondary tumors to integer
}
\details{
NSQIP encodes anyone with more than 8 secondary tumors as "8 or more".
This converts all "8 or more" to 8. If given NA, will return NA.
}
\examples{
nsqipr:::conv_hep_sec_numtumors(c("1","2","More than 8",NA))

}
\keyword{internal}
