% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{conv_inout}
\alias{conv_inout}
\title{Convert inout to logical}
\usage{
conv_inout(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
an integer vector
}
\description{
Convert inout to logical
}
\details{
If "Inpatient", will result in true. If given NA, will return NA.
}
\examples{

}
\keyword{internal}
