% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{conv_notno}
\alias{conv_notno}
\title{Convert "no" columns to logicals}
\usage{
conv_notno(vec)
}
\arguments{
\item{vec}{a character vector to convert "no" columns}
}
\value{
a logical vector
}
\description{
Convert "no" columns to logicals
}
\details{
Matches the case-insensitive fixed string \code{"no"} and \code{"none"} and negates the result.
Any string that does not match either of these will return TRUE. NA will return NA.
}
\examples{

x <- data.frame(x = rep("no", 10), y = rep("NO", 10), z = rep("yes", 10),
xx = rep("", 10), yy = rep(NA, 10), zz = rep("NONE", 10))
lapply(x, conv_notno)

}
\keyword{internal}
