% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{conv_numscale}
\alias{conv_numscale}
\title{Extract number scale from vector}
\usage{
conv_numscale(vec)
}
\arguments{
\item{vec}{a character vector to extract number scales from}
}
\value{
an integer vector
}
\description{
Extract number scale from vector
}
\details{
Matches the regex pattern \code{"^.*?\\d"}. NA will return NA.
}
\examples{

x <- data.frame(x = rep("1 - test", 10), y = rep("2- test", 10), z = rep("  3 test", 10),
xx = rep("4            5", 10), yy = rep(NA, 10))
lapply(x , nsqipr:::conv_numscale)

}
\keyword{internal}
