% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tar.R
\name{conv_open_assist}
\alias{conv_open_assist}
\title{Parse entries with open assist or hand assist}
\usage{
conv_open_assist(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries with open assist or hand assist
}
\details{
returns TRUE if either case-insensitive "w/ open assist" or
"hand assisted" is detected in the character vector. Cases with an
"open" or "open (planned)" approach will return NA.
}
\examples{
x <- c("Endoscopic","Endoscopic w/ open assist","Endoscopic w/ unplanned conversion to open",
"Hybrid","Hybrid w/ open assist","Hybrid w/ unplanned conversion to open",
"Laparoscopic","Laparoscopic w/ open assist","Laparoscopic w/ unplanned conversion to open",
"Laparoscopic hand assisted","NOTES","NOTES w/ open assist","NOTES w/ unplanned conversion to open",
"Open","Open (planned)","Other","Other MIS approach","Other MIS approach w/ open assist",
"Other MIS approach w/ unplanned conversion to open","Robotic","Robotic w/ open assist",
"Robotic w/ unplanned conversion to open","SILS","SILS w/ open assist",
"SILS w/ unplanned conversion to open", NA)

cbind(x, nsqipr:::conv_open_assist(x))

}
\keyword{internal}
