% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_pan.R
\name{conv_pan_cols}
\alias{conv_pan_cols}
\title{Convert targeted pancreatectomy columns}
\usage{
conv_pan_cols(df, filename)
}
\arguments{
\item{df}{a data table to be cleaned}

\item{filename}{the name of the file from which the data table has been read in}
}
\description{
Convert targeted pancreatectomy columns
}
\details{
If the file being processed is a targeted pancreatectomy data set,
it will be processed by this function. This function determines how data cleaning steps specific
to targeted pancreatectomy files should proceed.
}
\keyword{internal}
