% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_pan.R
\name{conv_pan_fistula}
\alias{conv_pan_fistula}
\title{Parse entries that indicate the presence of a fistula}
\usage{
conv_pan_fistula(vec)
}
\arguments{
\item{vec}{a character vector to parse}
}
\value{
a logical vector
}
\description{
Parse entries that indicate the presence of a fistula
}
\details{
returns TRUE if case-insensitive "yes" or "biochemical leak only'
is detected in the character vector.
}
\examples{
x <- c("No","No evidence of Biochemical Leak or POPF","Biochemical Leak only",
"Yes, Grade B POPF present","Yes, Grade C POPF present","Yes-clinical diagnosis, NPO-TPN",
"Yes-clinical diagnosis, drain continued >7 days",
"Yes-clinical diagnosis, percutaneous drainage performed",
"Yes-clinical diagnosis, reoperation performed",
"Yes-clinical diagnosis, spontaneous wound drainage","Yes-persistent drainage, NPO-TPN",
"Yes-persistent drainage, drain continued >7 days",
"Yes-persistent drainage, percutaneous drainage performed",
"Yes-persistent drainage, reoperation performed",
NA)

cbind(x, nsqipr:::conv_pan_fistula(x))

}
\keyword{internal}
