% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_pan.R
\name{conv_pan_fistula_intervention}
\alias{conv_pan_fistula_intervention}
\title{Parse a column for type of fistula intervention}
\usage{
conv_pan_fistula_intervention(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a factor vector
}
\description{
Parse a column for type of fistula intervention
}
\details{
NSQIP encodes the \code{pan_fistula} column as having one of multiple interventions.
This function extracts those values from character vectors and factors them.
}
\examples{
fistulas <- c("No", "Yes-persistent drainage, drain continued >7 days",
"Yes-clinical diagnosis, drain continued >7 days", "Yes-persistent drainage, percutaneous drainage performed",
"Yes-clinical diagnosis, percutaneous drainage performed", "Yes-persistent drainage, reoperation performed",
"Unknown", "Yes-clinical diagnosis, reoperation performed", "Yes-clinical diagnosis, spontaneous wound drainage",
"Yes-persistent drainage, NPO-TPN", "Yes-clinical diagnosis, NPO-TPN", "No evidence of Biochemical Leak or POPF",
"Biochemical Leak only", "Yes, Grade B POPF present", "Yes, Grade C POPF present")

nsqipr:::conv_pan_fistula_intervention(fistulas)

}
\keyword{internal}
