% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{conv_race}
\alias{conv_race}
\title{Convert race to factor}
\usage{
conv_race(vec, pacific = "asian")
}
\arguments{
\item{vec}{a character vector of races to be converted to a factor}

\item{pacific}{whether to consider "asian or pacific islander" as part of the "Native Hawaiian or Pacific islander"
level or part of the "Asian" level.}
}
\value{
a factor vector
}
\description{
Convert race to factor
}
\details{
2005-2007 NSQIP PUFs included a race called "Asian or Pacific islander". Later PUFs split these into "Asian" and
"Native Hawaiian or Pacific islander". In order to combine PUFs, a decision must be made as to which group to assign
"Asian or Pacific islander" to. To assign them to "Asian", \code{pacific} should be set to "asian". To assign them to
"Native Hawaiian or Pacific islander", \code{pacific} should be set to "hawaiian".
}
\examples{
x <- c("white","black or african american","asian or pacific islander")
conv_race(x)
x

x <- c("white","black or african american","asian or pacific islander")
conv_race(x, pacific = "hawaiian")
x
}
\keyword{internal}
