% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{conv_reasons}
\alias{conv_reasons}
\title{Converts readmission reasons to a factor}
\usage{
conv_reasons(vec)
}
\arguments{
\item{vec}{a character vector containing readmission reasons}
}
\value{
a factor vector
}
\description{
Converts readmission reasons to a factor
}
\details{
Formats factors accordingly:\tabular{lcr}{
   \bold{Original} \tab \bold{Level} \tab \bold{Label} \cr
   "superficial incisional ssi" \tab sssi \tab Superficial incisional SSI \cr
}
}
\examples{

x <- c("2000", "1900", "2020", "1950")
conv_date(x)

}
\keyword{internal}
