% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{conv_sex}
\alias{conv_sex}
\title{Convert sex to logical}
\usage{
conv_sex(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a logical vector
}
\description{
Convert sex to logical
}
\details{
if "male", will result in TRUE. If given NA, will return NA.
}
\examples{
 nsqipr:::conv_sex(c("male","MALE","female","FEMALE",NA))

}
\keyword{internal}
