% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_standard.R
\name{conv_to_standard}
\alias{conv_to_standard}
\title{The top level data cleaning function}
\usage{
conv_to_standard(file, cols, csv, rds)
}
\arguments{
\item{file}{a text file containing data to be cleaned}

\item{cols}{a character vector of column names}

\item{csv}{a character vector. write the resulting data table to a \code{.csv} file. See details for options.}

\item{rds}{a logical vector. Write the resulting data table to \code{.rds} file. Produces individual files for each \code{.txt} input.}
}
\description{
Calls all cleaning functions in the correct order and outputs a clean file.
}
\keyword{internal}
