% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{conv_yesno}
\alias{conv_yesno}
\title{Convert yes/no columns to logicals}
\usage{
conv_yesno(vec)
}
\arguments{
\item{vec}{a character vector to convert yes/no columns}
}
\value{
a logical vector
}
\description{
Convert yes/no columns to logicals
}
\details{
Matches the case-insensitive fixed string \code{"yes"}. NA will return NA.
}
\examples{

x <- data.frame(x = rep("yes", 10), y = rep("YES", 10), z = rep("no", 10),
xx = rep("NO", 10), yy = rep(NA, 10))
lapply(x , nsqipr:::conv_yesno)

}
\keyword{internal}
