% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_long.R
\name{convert_col_to_integer}
\alias{convert_col_to_integer}
\title{Convert a column to integers}
\usage{
convert_col_to_integer(df, col)
}
\arguments{
\item{df}{a data.table}

\item{col}{column to convert to integer}
}
\description{
Converts a column of character vectors to integer vectors.
}
\keyword{internal}
