% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-files.R
\name{create_dirs}
\alias{create_dirs}
\title{Creates directories from a vector of desired directory names. Only creates unique directories.}
\usage{
create_dirs(dirnames, base)
}
\arguments{
\item{dirnames}{a character vector of desired directory names}

\item{base}{the path of the containing directory}
}
\description{
Creates directories from a vector of desired directory names. Only creates unique directories.
}
\keyword{internal}
