% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_path}
\alias{create_path}
\title{Creates a path to a directory.}
\usage{
create_path(dir_name, base)
}
\arguments{
\item{dir_name}{desired name of directory}

\item{base}{the path of the containing directory}
}
\description{
Creates a path to a directory.
}
\keyword{internal}
