% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindo.R
\name{dindo}
\alias{dindo}
\title{Classifies a patient according to the Dindo-Clavien surgical complication grading scale.}
\usage{
dindo(df)
}
\arguments{
\item{df}{a dataframe including relevant columns containing information on specific post-operative complications.}
}
\value{
a numeric vector representing the Dindo-Clavien classification.
}
\description{
Classifies a patient according to the Dindo-Clavien surgical complication grading scale.
}
\keyword{internal}
