% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcon.R
\name{extract_exe_file}
\alias{extract_exe_file}
\title{Extract a file uzing \code{7z}}
\source{
\url{https://www.7-zip.org}
}
\usage{
extract_exe_file(file, tmpdir)
}
\arguments{
\item{file}{a file to be unzipped}

\item{tmpdir}{a temporary directory}
}
\description{
Unzips a file using the \code{7z} (\strong{7-zip}) command line interface.
}
\details{
\code{7z} is a file archiver with a high compression ratio format.
The command \code{e} "extracts files from archive (without using directory names).
The switch \code{-o} will "set output directory".
See \code{man 7z} for further details.
}
\keyword{internal}
