% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcon.R
\name{extract_exe_folder}
\alias{extract_exe_folder}
\title{Extract \code{.exe} files from a directory}
\usage{
extract_exe_folder(dir, tmpdir)
}
\arguments{
\item{dir}{A directory containing only \code{.exe} files downloaded directly from ACS NSQIP}

\item{tmpdir}{a temporary directory}
}
\description{
Accepts a directory and applys the function \code{\link{extract_exe_file}} to each \code{.exe} file.
}
\details{
The \code{tmpdir} parameter may be produced by the function \code{\link{build_tmp_dir}}.
}
\keyword{internal}
