% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{\%^\%}
\alias{\%^\%}
\title{Factor pipe}
\usage{
lhs \%^\% rhs
}
\arguments{
\item{lhs}{a character vector to be converted into a factor vector}

\item{rhs}{a named list specifying how to rename the levels}
}
\description{
A pipe that allows easy conversion of a vector into a factor with specified levels.
}
\examples{
nsqipr:::`\%^\%`(
c("apple","cherry","pork"),
list(fruit = c("apple", "banana", "cherry"), meat = c("steak", "chicken", "pork"))
)

}
\keyword{internal}
