% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcon.R
\name{get_file_or_dir}
\alias{get_file_or_dir}
\title{Check if a path is a directory or a file}
\usage{
get_file_or_dir(path, pattern)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link[base]{getwd}()}.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is performed.  Missing
    values will be ignored.  Elements with a marked encoding will
    be converted to the native encoding (and if that fails, considered
    non-existent).}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\value{
either a list of file paths or a single file path.
}
\description{
Checks if a provided path is a directory or a file. This allows a user to pass either a directory
of files matching a provided \code{pattern} or a single file matching a provided \code{pattern}.
}
\details{
\code{\link{get_file_or_dir}} is particularly necessary for \code{\link{extract_exe_folder}} because
\code{\link[base]{list.files}} will not accept a single file path and thus \code{\link{get_file_or_dir}}
is necssary to pass either a list of file paths or a single file path.
}
\keyword{internal}
