% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-readme.R
\name{get_nsqip_hospitals}
\alias{get_nsqip_hospitals}
\title{Fetch NSQIP hospital count}
\usage{
get_nsqip_hospitals()
}
\description{
This function scrapes the NSQIP participant website and returns the
total number of participating hospital.
}
\keyword{internal}
