% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-readme.R
\name{get_nsqip_pubmed}
\alias{get_nsqip_pubmed}
\title{Fetch NSQIP PubMed search}
\usage{
get_nsqip_pubmed()
}
\description{
This function queries PubMed for the search term 'NSQIP' and returns all the
results in a data frame. Note, it can only return up to 5000 results.
}
\keyword{internal}
