% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-readme.R
\name{get_nsqip_pubmed_journals}
\alias{get_nsqip_pubmed_journals}
\title{Count and prepare journals}
\usage{
get_nsqip_pubmed_journals(pubmed, num = 10)
}
\arguments{
\item{pubmed}{A PubMed search data frame}

\item{num}{number of top journals to keep in data frame}
}
\description{
This function groups by and counts journal names. It selects the top \code{num}
number of journals. It additionally calculates an angle and horizontal adjustment
necessary for use in a radial bar plot.
}
\keyword{internal}
