% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-readme.R
\name{get_nsqip_pubmed_yearsums}
\alias{get_nsqip_pubmed_yearsums}
\title{Calculate NSQIP search hits cumulative sum}
\usage{
get_nsqip_pubmed_yearsums(pubmed, interval = 5)
}
\arguments{
\item{pubmed}{a PubMed search data frame. Must have a year column.}

\item{interval}{filter for the years divisible by this interval.}
}
\description{
This function takes a data frame of search hits from a PubMed query and
calculates the cumulative sum of hits per year. Meant to be used in conjunction
with a data frame returned from the \code{get_nsqip_pubmed} function.
}
\keyword{internal}
