% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{get_pufyear}
\alias{get_pufyear}
\title{Add a PUF year column}
\usage{
get_pufyear(df, filename)
}
\arguments{
\item{df}{a data table to append the new column to}

\item{filename}{the file name from which the record was derived}
}
\value{
a data table with a new ordered factor column called \code{pufyear}
}
\description{
This column notes the file from which the record came.
}
\details{
Matches the year of the file from the file name using a regular expression.
Requires that the original file names \bold{are not changed}.
}
\examples{
data.table::data.table(x = rep("name", 10))
get_pufyear(x, "acs_nsqip_puf12.txt")
x$pufyear < "2013"
x$pufyear > "2005-2006"

}
\keyword{internal}
