% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%qsin\%}
\alias{\%qsin\%}
\title{Quick Search}
\usage{
str \%qsin\% vec
}
\arguments{
\item{str}{a character vector: the values to be matched}

\item{vec}{a character vector: the values to be matched against}
}
\value{
a logical vector of length \code{length(str)}
}
\description{
Quick case-insensitive search of strings in a character vector
}
\details{
Utilizes \code{data.table::\%chin\%} to rapidly complete a case-insensitive search
through a character vector to return a logical vector of string detections.
Will always return TRUE or FALSE for each position of \code{str} regardless of NA missing values
in either provided vector. NA in \code{str} will never match an NA value in \code{vec}.
}
\examples{
x <- c("apple","banana","cherry",NA)
nsqipr:::`\%qsin\%`("apple",x) # could also use "apple" \%qsin\% x
nsqipr:::`\%qsin\%`(c("APPLE","BANANA","coconut", NA), x)

}
