% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{insulin}
\alias{insulin}
\title{Parse a column for insulin usage}
\usage{
insulin(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a logical vector
}
\description{
Parse a column for insulin usage
}
\details{
NSQIP encodes the \code{diabetes} column as either "no", "non-insulin", or "insulin".
This function checks that the value is both \bold{not} "no" and also equal to "insulin". Returns
FALSE if either "no" or "non-insulin". If given NA, will return NA.
}
\examples{

}
\keyword{internal}
