% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindo.R
\name{isDead}
\alias{isDead}
\title{Check if a patient has died within 30 days of the index procedure. Assumes that if a patient died, a date would be
recorded in \code{yrdeath}.}
\usage{
isDead(col)
}
\arguments{
\item{col}{a vector of any type}
}
\value{
a logical vector. TRUE if not NA, FALSE if NA.
}
\description{
Check if a patient has died within 30 days of the index procedure. Assumes that if a patient died, a date would be
recorded in \code{yrdeath}.
}
\keyword{internal}
