% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isFullDT}
\alias{isFullDT}
\title{Check if a data table is full}
\usage{
isFullDT(x)
}
\arguments{
\item{x}{a data.table}
}
\description{
Check if a data table is full
}
\details{
Will return TRUE if the object supplied is a non-empty data.table.
Otherwise, returns FALSE.
}
\examples{
isFullDT(NULL)
isFullDT("test")
isFullDT(NA)
isFullDT(data.table::data.table())
isFullDT(data.table::data.table(test = numeric(), test2 = logical(), test3 = character()))
isFullDT(data.table::data.table(test = 1, test2 = TRUE, test3 = "character"))

}
