% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{make_anesthes_other_cols}
\alias{make_anesthes_other_cols}
\title{Create reoperation columns for long conversion}
\usage{
make_anesthes_other_cols(df)
}
\arguments{
\item{df}{a data table to add the columns to}
}
\description{
Create reoperation columns for long conversion
}
\details{
First checks if the data table contains an "anesthes_other" column.
If so, the column is split into 8 columns according to the regex pattern ",\\s?".
}
\examples{
x <- data.table::data.table(anesthes_other = c("General","General, Spinal", "General, Spinal, MAC/IV Sedation", NA))
make_anesthes_other_cols(x)
x

}
\keyword{internal}
