% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_long.R
\name{make_cols_long}
\alias{make_cols_long}
\title{Convert columns to a wide format}
\usage{
make_cols_long(df, ..., variable.name, na.cols, removeFALSE, reorder = FALSE)
}
\arguments{
\item{df}{a data.table containing the columns in a wide format}

\item{...}{character vectors of column names that will be melted together into single columns.}

\item{variable.name}{desired output variable name (optional)}

\item{na.cols}{character vector of column names indicating which records
should be removed when NA values are present in these columns (optional).}

\item{removeFALSE}{character vector of a column name indicating which records
should be removed when FALSE values are present in this column (optional).}

\item{reorder}{logical vector indicating whether records should be renumbered after removal
of NA and/or FALSE records.}
}
\value{
data.table
}
\description{
This is a complex function that accepts specified lists of columns and melts them
into a long format from a wide format. This is designed for those columns in the original
NSQIP dataset that are stored in a wide format, such as the CPT, readmission and
reoperation columns; this is not meant for columns that store data in a comma-separated
fashion (such as \code{anesthes_other}).
}
\details{
The \code{...} parameter accepts multiple character vectors that should each contain
the names of columns that will be melted together to create a single column in the long format
output. For example, a character vector named "reoperation" will contain the names "reoperation1",
"reoperation2", and "reoperation3", indicating these three columns in the original dataset will
be melted into a single column called "reoperation" in the long format.

The \code{removeFALSE} character vector should only specify a single column.
}
\seealso{
make_commas_long
}
\keyword{internal}
