% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.R
\name{make_hep_con_ablation_long}
\alias{make_hep_con_ablation_long}
\title{Convert hep_con_ablation_140101 column from wide to long format}
\usage{
make_hep_con_ablation_long(df)
}
\arguments{
\item{df}{a data.table}
}
\value{
a data.table
}
\description{
Convert hep_con_ablation_140101 column from wide to long format
}
\details{
The data contained in the 5 "hep_con_ablation" columns created by \code{make_hep_con_ablation_cols}
are converted into a long format.

If "hep_con_ablation_140101" is a column in \code{df}, it will be broken into a long format with
\code{caseid} as the ID variable for joining back to the main table. This is because the targeted hepatectomy
datasets input multiple values into a single "hep_con_ablation_140101" column. For example,
"Microwave ablation,Other ablation" may be an entry in the raw data set. This makes
parsing for patients that had other types of concurrent ablation at any point, for example, very difficult.

Note that this does not alter the "hep_con_ablation_140101" column.
}
\examples{
x <- data.table::data.table(caseid = 1:12,
hep_con_ablation_140101 = c("Microwave ablation", "RFA ablation", "Other ablation",
"Microwave ablation,Other ablation", "RFA ablation,Microwave ablation",
"RFA ablation,Other ablation", "RFA ablation,Alcohol ablation", "Alcohol ablation",
"Microwave ablation,Alcohol ablation", "Cryoablation", NA, "RFA ablation,Cryoablation"))

nsqipr:::make_hep_con_ablation_cols(x)
nsqipr:::make_hep_con_ablation_long(x)

}
\keyword{internal}
