% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.R
\name{make_hep_invasive_type_long}
\alias{make_hep_invasive_type_long}
\title{Convert hep_invasive_type column from wide to long format}
\usage{
make_hep_invasive_type_long(df)
}
\arguments{
\item{df}{a data.table}
}
\value{
a data.table
}
\description{
Convert hep_invasive_type column from wide to long format
}
\details{
The data contained in the 5 "hep_invasive_type" columns created by \code{make_hep_invasive_type_cols}
are converted into a long format.

If "hep_invasive_type" is a column in \code{df}, it will be broken into a long format with
\code{caseid} as the ID variable for joining back to the main table. This is because the targeted hepatectomy
datasets input multiple values into a single "hep_invasive_type" column. For example,
"Biliary stent for biliary obstruction/leak,Other intervention" may be an entry in the raw data set. This makes
parsing for patients that had other types of invasive procedures at any point, for example, very difficult.

Note that this does not alter the "hep_invasive_type" column.
}
\examples{
x <- data.table::data.table(caseid = 1:7,
hep_invasive_type = c("Biliary stent for biliary obstruction/leak", "Bilirubin-rich fluid from drain or aspirate", "Other intervention",
"Biliary stent for biliary obstruction/leak,Other intervention", "Bilirubin-rich fluid from drain or aspirate,Biliary stent for biliary obstruction/leak",
"Bilirubin-rich fluid from drain or aspirate,Other intervention", NA))

hep_invasive_type_cols <- paste("hep_invasive_type", 1:5, sep = "")

nsqipr:::make_long_cols(x, "hep_invasive_type", hep_invasive_type_cols)
nsqipr:::make_hep_invasive_type_long(x)

}
\keyword{internal}
