% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-readme.R
\name{make_nsqip_hits_graph}
\alias{make_nsqip_hits_graph}
\title{Make a graph of cumulative PubMed search hits}
\usage{
make_nsqip_hits_graph(pubmed, yearsums)
}
\arguments{
\item{pubmed}{a PubMed search data frame. Must have a date and n column.}

\item{yearsums}{a data frame of cumulative sums. Used to add points at
specific intervals to the line.}
}
\description{
This graph will chart the cumulative sum of PubMed search hits.
Meant to be used with the \code{get_nsqip_pubmed} and \code{get_nsqip_pubmed_yearsums}
functions.
}
\keyword{internal}
