% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-readme.R
\name{make_nsqip_journal_circular}
\alias{make_nsqip_journal_circular}
\title{Create circular bar plot of journal hits}
\usage{
make_nsqip_journal_circular(journals)
}
\arguments{
\item{journals}{a data frame of journals and number of hits. Should also have
parameters for angles and horizontal adjustments for display in a circular plot}
}
\description{
This function creates a circular bar plot that displays the number of hits
from a PubMed search that are published in each journal.
}
\keyword{internal}
