% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_pan.R
\name{make_pan_percdrainage_cols}
\alias{make_pan_percdrainage_cols}
\title{Create percutaneous drainage columns for long conversion}
\usage{
make_pan_percdrainage_cols(df)
}
\arguments{
\item{df}{a data table to add the columns to}
}
\description{
Create percutaneous drainage columns for long conversion
}
\details{
First checks if the data table contains a "pan_percdrainage" column.
If so, the column is split into 4 columns according to the regex pattern ",\\s?".
}
\examples{
x <- data.table::data.table(
pan_percdrainage = c("Yes-other", "Yes-bile", "Yes-pus", "Yes-amylase-rich fluid",
"Yes-amylase-rich fluid,Yes-pus", "Yes-bile,Yes-other", "Yes-amylase-rich fluid,Yes-pus,Yes-other",
"Yes-amylase-rich fluid,Yes-bile,Yes-other", "Yes-pus,Yes-bile", "Yes-pus,Yes-other",
"Yes-amylase-rich fluid,Yes-pus,Yes-bile", "Yes-amylase-rich fluid,Yes-other",
"Yes-amylase-rich fluid,Yes-pus,Yes-bile,Yes-other", "Yes-amylase-rich fluid,Yes-bile",
"Yes-pus,Yes-bile,Yes-other")
)

nsqipr:::make_pan_percdrainage_cols(x)
x

}
\keyword{internal}
