% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.R
\name{make_pan_percdrainage_long}
\alias{make_pan_percdrainage_long}
\title{Convert pan_percdrainage column from wide to long format}
\usage{
make_pan_percdrainage_long(df)
}
\arguments{
\item{df}{a data.table}
}
\value{
a data.table
}
\description{
Convert pan_percdrainage column from wide to long format
}
\details{
The data contained in the 4 "pan_percdrainage" columns created by \code{make_pan_percdrainage_cols}
are converted into a long format.

If "pan_percdrainager" is a column in \code{df}, it will be broken into a long format with
\code{caseid} as the ID variable for joining back to the main table. This is because the targeted pancreatectomy
datasets input multiple values into a single "pan_percdrainage" column. For example,
"Yes-other, Yes-bile, Yes-pus" may be an entry in the raw data set. This makes
parsing for patients that had purulent percutaneous drainage at any point, for example, very difficult.

Note that this does not alter the "pan_percdrainage" column.
}
\examples{
x <- data.table::data.table(caseid = 1:15,
pan_percdrainage = c("Yes-other", "Yes-bile", "Yes-pus", "Yes-amylase-rich fluid",
"Yes-amylase-rich fluid,Yes-pus", "Yes-bile,Yes-other", "Yes-amylase-rich fluid,Yes-pus,Yes-other",
"Yes-amylase-rich fluid,Yes-bile,Yes-other", "Yes-pus,Yes-bile", "Yes-pus,Yes-other",
"Yes-amylase-rich fluid,Yes-pus,Yes-bile", "Yes-amylase-rich fluid,Yes-other",
"Yes-amylase-rich fluid,Yes-pus,Yes-bile,Yes-other", "Yes-amylase-rich fluid,Yes-bile",
"Yes-pus,Yes-bile,Yes-other"))

nsqipr:::make_pan_percdrainage_cols(x)
nsqipr:::make_pan_percdrainage_long(x)

}
\keyword{internal}
