% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.R
\name{make_readm_long}
\alias{make_readm_long}
\title{Convert readmission columns from wide to long format}
\usage{
make_readm_long(df, removeFALSE = FALSE)
}
\arguments{
\item{df}{a data.table}

\item{removeFALSE}{a logical indicating whether to remove rows where \code{reoperation} is FALSE}
}
\value{
a data.table
}
\description{
Convert readmission columns from wide to long format
}
\details{
The data from the data table is melted into a long format with \code{caseid} as the ID variable to allow
rejoining to the main table. After melting, rows with missing values are omitted to reduce the size of the table.
Rows where \code{readmission} are false may also be removed with \code{removeFALSE} to reduce table size if
desired, but note this results in an inability to a clarify a known FALSE ("did not have a readmission") from
a missing value ("do not know if there was a readmission").
}
\examples{

}
\keyword{internal}
