% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_file}
\alias{move_file}
\title{Moves files from a parent directory to newly created subdirectories.}
\usage{
move_file(dirnames, base)
}
\arguments{
\item{dirnames}{a character vector of desired directory names}

\item{base}{the path of the containing directory}
}
\description{
Moves files from a parent directory to newly created subdirectories.
}
\keyword{internal}
