% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_files}
\alias{move_files}
\title{Moves files from a parent directory to newly created subdirectories.}
\usage{
move_files(files)
}
\arguments{
\item{files}{character vector of files to move}
}
\description{
Moves files from a parent directory to newly created subdirectories.
}
\keyword{internal}
