% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_standard.R
\name{nsqip_dir}
\alias{nsqip_dir}
\title{Runs the \code{nsqip} function in a single directory.}
\usage{
nsqip_dir(dir, rds, csv, dataframe)
}
\arguments{
\item{dir}{directory to which the \code{nsqip} function is being applied.}
}
\description{
Utilized by \code{nsqip} after all files are parsed and respective directories are created.
}
\keyword{internal}
