% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsqipr.R
\docType{package}
\name{nsqipr}
\alias{nsqipr}
\title{nsqipr: A cleaner NSQIP. \if{html}{\figure{hex.png}{options: align="right" alt='nsqipr logo' width=200}} \if{latex}{\figure{hex.png}{options: width=0.5in}}}
\description{
The purpose of \code{nsqipr} is to streamline the data cleaning process.
This package is geared towards those surgical interns, residents, and attendings
who have limited experience with R, SQL, or "big data" analysis.
It is also designed to be a useful tool for that experienced researcher or
computer scientist making frequent use of ACS NSQIP PUFs.
}
\seealso{
\itemize{
\item{\href{https://github.com/dylanrussellmd/nsqipr}{nsqipr Github}}
\item{\href{https://www.dylanrussellmd/nsqipr_book}{nsqipr Design and Data Dictionary}}
\item{\href{https://www.dylanrussellmd/nsqipr}{nsqipr documentation}}
\item{\href{https://www.facs.org/quality-programs/acs-nsqip}{ACS NSQIP}}
}
}
\author{
Dylan Russell, M.D.
}
