% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_dir}
\alias{open_dir}
\title{Opens a directory and lists all \code{.txt} files with full names.}
\usage{
open_dir(path)
}
\arguments{
\item{path}{a directory path}
}
\description{
Opens a directory and lists all \code{.txt} files with full names.
}
\keyword{internal}
