% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-files.R
\name{parse_filename}
\alias{parse_filename}
\title{Parses a file name and returns a lower-case string of the extracted regular expression match.}
\usage{
parse_filename(file)
}
\arguments{
\item{file}{a file name}
}
\value{
string
}
\description{
Parses a file name and returns a lower-case string of the extracted regular expression match.
}
\keyword{internal}
