% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-files.R
\name{parse_files}
\alias{parse_files}
\title{Parses a directory of files and creates the requisite directories. Moves files into their respective directories.}
\usage{
parse_files(files)
}
\arguments{
\item{files}{a character vector of file names.}
}
\description{
Parses a directory of files and creates the requisite directories. Moves files into their respective directories.
}
\keyword{internal}
