% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb.R
\name{pb}
\alias{pb}
\title{Create a progress bar}
\usage{
pb(write_to_csv)
}
\arguments{
\item{write_to_csv}{boolean value indicating if the user would like to write to a CSV.}
}
\description{
This function creates a progress bar for use with the \code{nsqip} function.
}
\keyword{internal}
