% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-readme.R
\name{prep_title}
\alias{prep_title}
\title{Prepare journal titles}
\usage{
prep_title(str)
}
\arguments{
\item{str}{a journal name}
}
\description{
This function will strip extraneous characters and subtitles from a journal name.
}
\keyword{internal}
