% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{remove_undesired}
\alias{remove_undesired}
\title{Remove undesired columns from a data table}
\usage{
remove_undesired(df, undesired_cols)
}
\arguments{
\item{df}{a data table from which to remove columns}

\item{undesired_cols}{a character vector of column names identifying columns to be removed}
}
\value{
a data table
}
\description{
Remove undesired columns from a data table
}
\details{
This function \bold{modifies by reference}.
}
\examples{
x <- data.table::data.table(a = c(1,2,3), b = c(1,2,3), c = c(1,2,3))
orignames <- names(x)
undesired_cols <- c("a","b", "d")
remove_undesired(x, undesired_cols)
identical(names(x), setdiff(orignames, undesired_cols))

}
\keyword{internal}
