% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{setlower}
\alias{setlower}
\title{Convert all variables to lower case}
\usage{
setlower(df)
}
\arguments{
\item{df}{a data table to convert to lower case}
}
\value{
a data table
}
\description{
Convert all variables to lower case
}
\details{
This function \bold{modifies by reference}.
}
\examples{

x <- data.table::data.table(x = rep("APPLE", 10), y = rep("BANANA", 10), z = rep("CHERRY", 10))
setlower(x)
x

}
\keyword{internal}
