% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{setlowernames}
\alias{setlowernames}
\title{Convert column names to lower case}
\usage{
setlowernames(df)
}
\arguments{
\item{df}{a data frame to convert names to lower case}
}
\value{
a data frame
}
\description{
Convert column names to lower case
}
\examples{

x <- data.frame(X = NA, Y = NA, Z = NA)
setlowernames(x)
names(x)

}
\keyword{internal}
