% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{setna}
\alias{setna}
\title{Convert all strings matching vector to NA}
\usage{
setna(df, val)
}
\arguments{
\item{df}{a data table to convert values to NA}

\item{val}{a character vector of values to set to NA}
}
\value{
a data table
}
\description{
Convert all strings matching vector to NA
}
\details{
This function \bold{modifies by reference}.
}
\examples{

xx <- data.table::data.table(x = rep("unknown", 10), y = rep("unknown/not reported", 10), z = rep("null", 10),
xx = rep("not documented", 10), yy = rep("none/not documented", 10), zz = rep("not entered", 10),
xxx = rep("-99", 10), yyy = rep("test", 10))

setna(x, c("unknown", "unknown/not reported", "null", "n/a", "not documented", "none/not documented", "not entered","-99"))
x

}
\keyword{internal}
