% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb.R
\name{tick}
\alias{tick}
\title{Increment a progress bar}
\usage{
tick(pb, preface, file, len = 1)
}
\arguments{
\item{pb}{a progress bar object}

\item{preface}{the desired preface text to a progress bar}

\item{file}{the file being processed}

\item{len}{number of ticks}
}
\description{
Increments a progress bar by a specified number of ticks. Meant to be used in a pipe.
}
\keyword{internal}
