% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{type_prsepis}
\alias{type_prsepis}
\title{Parse a column for type of sepsis}
\usage{
type_prsepis(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a factor vector
}
\description{
Parse a column for type of sepsis
}
\details{
NSQIP encodes the \code{sepsis} column as either "sirs", "sepsis", "septic shock", or "none.
This function factors the vector for the levels "SIRS", "Sepsis", and "Septic shock".

\bold{NOTE}: \code{prsepis} is spelled illogically (as it is originally spelled in the NSQIP database).
It is not spelled \code{prsepsis}.
}
\examples{
 nsqipr:::type_prsepis(c("sirs","sepsis", "septic shock", NA))

}
\keyword{internal}
