% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-readme.R
\name{unescape_xml}
\alias{unescape_xml}
\title{Translate XML}
\usage{
unescape_xml(str)
}
\arguments{
\item{str}{an XML string}
}
\description{
PubMed searches are returned in XML format. This will convert special characters
(such as &) into the correct characters for display.
}
\keyword{internal}
