% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcon.R
\name{write_from_file}
\alias{write_from_file}
\alias{parse_filename}
\alias{create_df}
\alias{check_table}
\alias{write_to_table}
\title{Enact all necessary steps to upload delimited text file to a database}
\usage{
write_from_file(file, conn)

parse_filename(file)

create_df(file)

check_table(conn, tablename, df)

write_to_table(tablename, df)
}
\arguments{
\item{file}{a delimited text file to be written to a database}

\item{conn}{An S4 object of class PostgreSQL}

\item{tablename}{the name of the table}

\item{df}{a dataframe}
}
\description{
Determines to which table to write (\code{\link{parse_filename}} and \code{\link{check_table}}),
creates a data frame from \code{file} (\code{\link{create_df}}), and writes data to the specified
database table (\code{\link{write_to_table}}.
}
\details{
The temporary \code{.csv} that is created is immediately destroyed after completion of the function in
order to excessive storage requirements.

A \strong{URI} is created with the following format in order to access the database via a system command:
\code{postgresql://NSQIP_DB_USER:NSQIP_DB_PW@NSQIP_DB_HOST:NSQIP_DB_PORT/NSQIP_DB}. This requires these
variables to be declared in the local \code{.Renviron} file.

The system command to write to the database requires that \code{psql} is installed on the host system
and available in the PATH.
}
\section{Functions}{
\itemize{
\item \code{parse_filename}: parses a file name with the regular expression
"\code{acs_nsqip_puf|puf_tar_[a-z]{1,4}}". Returns a lower case table name -
either \code{acs_nsqip_puf} for all ACS NSQIP PUF files or \code{puf_tar_x}
where \code{x} is a three- or four-letter abbreviation for the targeted procedure.

\item \code{create_df}: creates a dataframe from the delimited text file passed by \code{file}.
Will eventually be separated into a separate \code{.R} file with multiple functions to provide
more sophisticated formatting.

\item \code{check_table}: checks if the database contains a table equal to \code{tablename}.
If it does not, it creates the table using the variable names specified in the provided dataframe.

\item \code{write_to_table}: creates a temporary \code{.csv} file from a dataframe and writes the data
to a PostgreSQL 12 database via a system command.
}}

\keyword{internal}
