% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcon.R
\name{write_to_database}
\alias{write_to_database}
\title{Write a delimited text file to a database}
\usage{
write_to_database(conn, tmpdir)
}
\arguments{
\item{conn}{An S4 object of class PostgreSQL}

\item{tmpdir}{a temporary directory}
}
\description{
Accepts a directory containing \code{.txt} files to be written to a database
and applies the function \code{\link{write_from_file}} to each file.
}
\keyword{internal}
