% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_app.R
\name{conv_app_intraabscess_intervention}
\alias{conv_app_intraabscess_intervention}
\title{Parse a column for type of intra-abdominal abscess intervention}
\usage{
conv_app_intraabscess_intervention(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a factor vector
}
\description{
Parse a column for type of intra-abdominal abscess intervention
}
\details{
NSQIP encodes the \code{app_intraabscess} column as having one of multiple interventions.
This function extracts those values from character vectors and factors them.
}
\examples{

}
\keyword{internal}
