% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{conv_hispanic}
\alias{conv_hispanic}
\alias{conv_hispanic_helper}
\title{Add or update hispanic ethnicity column}
\usage{
conv_hispanic(df)

conv_hispanic_helper(df)
}
\arguments{
\item{df}{a data.table to add to or update with an \code{ethnicity_hispanic} column}

\item{race}{a character vector containing race}

\item{ethnicity_hispanic}{a character vector containing "yes" and "no" values regarding hispanic ethnicity}
}
\value{
a data table
}
\description{
Add or update hispanic ethnicity column
}
\details{
\code{ethnicity_hispanic} was not added until the 2008 NSQIP PUF when \code{race} was revised to
\code{race_new}. Data regarding hispanic ethnicity was hard coded directly into the old \code{race} variable
(such as "Hispanic, White"). In order to marry early and later datasets, this information must be extracted
from \code{race} and a new \code{ethnicity_hispanic} column created.

If the data provided already has a \code{ethnicity_hispanic} column present, this column is simply converted
into a logical vector.
}
\section{Functions}{
\itemize{
\item \code{conv_hispanic_helper}: A helper function for updating the \code{ethnicity_hispanic} column
}}

\examples{
x <- data.table::data.table(race = c("Hispanic, White", "White, Not of Hispanic Origin","Hispanic, Black","Black, Not of Hispanic Origin",
"American Indian or Alaska Native","Asian","Native Hawaiian or Pacific Islander","Asian or Pacific Islander"))
conv_hispanic(x)
x

x <- data.table::data.table(race = c("Hispanic, White", "White, Not of Hispanic Origin","White","Hispanic, Black", "Black, Not of Hispanic Origin","Black or African American",
+ "American Indian or Alaska Native","Asian","Native Hawaiian or Pacific Islander","Asian or Pacific Islander"),
+ ethnicity_hispanic = c(NA, NA, "yes", NA, NA, NA, NA, NA, NA, NA))
conv_hispanic(x)
x

}
\keyword{internal}
