% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-files.R
\name{get_file_or_dir}
\alias{get_file_or_dir}
\title{Check if a path is a directory or a file}
\usage{
get_file_or_dir(
  path,
  pattern = "^(acs_nsqip_puf|puf_tar_[a-z]{1,4})(?:\\\\d{2})?(?:.*)?\\\\.txt$"
)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link[base]{getwd}()}.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is performed.  Missing
    values will be ignored.  Elements with a marked encoding will
    be converted to the native encoding (and if that fails, considered
    non-existent).}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\value{
either a list of file paths or a single file path.
}
\description{
Checks if a provided path is a directory or a file. This allows a user to pass either a directory
of files matching a provided \code{pattern} or a single file matching a provided \code{pattern}.
}
\keyword{internal}
