% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isFullDT}
\alias{isFullDT}
\title{Check if a data table is full}
\usage{
isFullDT(x)
}
\arguments{
\item{x}{a data.table}
}
\description{
Check if a data table is full
}
\details{
Will return TRUE if the object supplied is a non-empty data.table.
Otherwise, returns FALSE.
}
\examples{
nsqipr:::isFullDT(NULL)
nsqipr:::isFullDT("test")
nsqipr:::isFullDT(NA)
nsqipr:::isFullDT(data.table::data.table())
nsqipr:::isFullDT(data.table::data.table(test = numeric(), test2 = logical(), test3 = character()))
nsqipr:::isFullDT(data.table::data.table(test = 1, test2 = TRUE, test3 = "character"))

}
