% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsqip.R
\name{nsqip}
\alias{nsqip}
\title{Clean NSQIP data}
\usage{
nsqip(path, rds = TRUE, csv = TRUE, dataframe = FALSE)
}
\arguments{
\item{path}{path to file or directory.}

\item{return_df}{return a data frame.}

\item{write_to_csv}{write the resulting data frame to a \code{.csv} file. Produces individual files for each \code{.txt} input}

\item{append}{writes the resulting data frames all to a single \code{.csv} file}

\item{headers}{whether to include column headers in the generated \code{.csv} file}
}
\description{
Converts a single NSQIP \code{.txt} or a directory of NSQIP \code{.txt} files into a standardized NSQIP data frame.
}
