% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_standard.R
\name{nsqip}
\alias{nsqip}
\title{Converts a single NSQIP \code{.txt} or a directory of NSQIP \code{.txt} files into a standardized NSQIP data frame.}
\usage{
nsqip(path, write_to_csv = FALSE, return_df = TRUE)
}
\arguments{
\item{path}{path to file or directory.}

\item{write_to_csv}{write the resulting data frame to a \code{.csv} file}

\item{return_df}{return the resulting data frame}
}
\value{
a data frame of class \code{tibble}.
}
\description{
Converts a single NSQIP \code{.txt} or a directory of NSQIP \code{.txt} files into a standardized NSQIP data frame.
}
