% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcon.R
\name{nsqip}
\alias{nsqip}
\title{Write data from ACS NSQIP files into database.}
\source{
\url{https://www.facs.org/quality-programs/acs-nsqip}
}
\usage{
nsqip(dir, conn)
}
\arguments{
\item{dir}{A directory containing only \code{.exe} files downloaded directly from ACS NSQIP.}

\item{conn}{An S4 object of class PostgreSQL}
}
\description{
Accepts a directory containing ACS NSQIP files as downloaded from the ACS NSQIP website and uploads to a database.
Will automatically format all strings in sentence case and correctly set the variable types.
Table names will be either \code{acs_nsqip_puf} for all ACS NSQIP PUF files or
\code{puf_tar_x} where \code{x} is a three- or four-letter abbreviation for the targeted procedure.
}
\section{Warning}{
\itemize{
\item The directory should include \strong{only} those ".exe" executable files downloaded directly from ACS NSQIP.
\item Do \strong{not} change the names of the downloaded files.
\item Currently only functioning with a PostgreSQL 12 database server.
\item Requires \code{7z} and \code{psql} installed on the host OS. Be sure \code{psql} is in the PATH.
}
}

\examples{
\dontrun{
nsqip(dir, conn)
DBI::dbGetQuery(conn, "select count(*) as n from acs_nsqip_puf")
}

}
