% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsqip.R
\name{nsqip}
\alias{nsqip}
\title{Clean NSQIP data}
\usage{
nsqip(path, rds = TRUE, csv = TRUE, datatable = FALSE)
}
\arguments{
\item{path}{path to file or directory.}

\item{rds}{write the resulting data frame to \code{.rds} file. Produces individual files for each \code{.txt} input}

\item{csv}{write the resulting data frame to a \code{.csv} file. Produces individual files for each \code{.txt} input}

\item{datatable}{return a data table?}
}
\value{
if \code{datatable} is TRUE, will return a data.table
}
\description{
Converts a single NSQIP \code{.txt} or a directory of NSQIP \code{.txt} files into a standardized NSQIP data table.
}
