% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_standard.R
\name{nsqip_dir}
\alias{nsqip_dir}
\title{Runs the \code{nsqip} function in a single directory.}
\usage{
nsqip_dir(dir, csv, rds)
}
\arguments{
\item{dir}{directory to which the \code{nsqip} function is being applied.}

\item{csv}{write the resulting data frame to a \code{.csv} file. See details for options.}

\item{rds}{a logical value. Write the resulting data frame to \code{.rds} file. Produces individual files for each \code{.txt} input.}
}
\description{
Utilized by \code{nsqip} after all files are parsed and respective directories are created.
}
\keyword{internal}
