% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_standard.R
\name{nsqip_dir}
\alias{nsqip_dir}
\title{Runs the \code{nsqip} function in a single directory.}
\usage{
nsqip_dir(dir, return_df, write_to_csv, append, headers)
}
\arguments{
\item{dir}{directory to which the \code{nsqip} function is being applied.}

\item{return_df}{return a data frame.}

\item{write_to_csv}{write the resulting data frame to a \code{.csv} file. Produces individual files for each \code{.txt} input}

\item{append}{writes the resulting data frames all to a single \code{.csv} file}

\item{headers}{whether to include column headers in the generated \code{.csv} file}
}
\description{
Utilized by \code{nsqip} after all files are parsed and respective directories are created.
}
\keyword{internal}
