% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb.R
\name{pb}
\alias{pb}
\title{Create a progress bar}
\usage{
pb(csv, rds)
}
\arguments{
\item{csv}{character vector indicating if the user would like to write to a CSV.}

\item{rds}{logical vector indicating if the user would like to write an RDS.}
}
\description{
This function creates a progress bar for use with the \code{nsqip} function.
}
\keyword{internal}
