% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{setlower}
\alias{setlower}
\title{Convert all variables to lower case}
\usage{
setlower(df)
}
\arguments{
\item{df}{a data frame to convert to lower case}
}
\value{
a data frame
}
\description{
Convert all variables to lower case
}
\examples{

x <- data.frame(x = rep("APPLE", 10), y = rep("BANANA", 10), z = rep("CHERRY", 10))
setlower(x)
x

}
\keyword{internal}
