% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conv.R
\name{setna}
\alias{setna}
\title{Convert all strings matching vector to NA}
\usage{
setna(df, val)
}
\arguments{
\item{df}{a data table to convert values to NA}

\item{val}{a character vector of values to set to NA}
}
\value{
a data table
}
\description{
Convert all strings matching vector to NA
}
\details{
This function \bold{modifies by reference}.
}
\examples{

x <- data.table::data.table(x = c(rep("unknown", 10), rep("test", 10)), y = c(rep("unknown/not reported", 10), rep("test", 10)), z = c(rep("null", 10), rep("test", 10)),
xx = c(rep("not documented", 10), rep("test", 10)), yy = c(rep("none/not documented", 10), rep("test", 10)), zz = c(rep("not entered", 10), rep("test", 10)),
xxx = c(rep("-99", 10), rep("test", 10)), yyy = c(rep("test", 10), rep("test", 10)))

setna(x, c("unknown", "unknown/not reported", "null", "n/a", "not documented", "none/not documented", "not entered","-99"))
x

}
\keyword{internal}
