% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_acs.R
\name{when_dyspnea}
\alias{when_dyspnea}
\title{Parse a column for type of dyspnea}
\usage{
when_dyspnea(vec)
}
\arguments{
\item{vec}{a character vector of values to convert}
}
\value{
a logical vector
}
\description{
Parse a column for type of dyspnea
}
\details{
NSQIP encodes the \code{dyspnea} column as either "no", "at rest", or "moderate exertion".
This function checks that the value is both \bold{not} "no" and also equal to "insulin". Returns
FALSE if either "no" or "non-insulin". If given NA, will return NA.
}
\examples{

}
\keyword{internal}
